+++ test/db/models.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
 class User < ActiveRecord::Base
-  has_one :profile
+  has_many :profiles
+  has_one :default_profile, -> { where(default: true) }, class_name: 'Profile'
   belongs_to :forum
 
   has_many :questions, foreign_key: :author_id
   has_many :answers,   foreign_key: :author_id
   has_many :votes
+  has_many :reactions
 
   has_and_belongs_to_many :roles, join_table: :user_roles
 end
@@ -37,6 +39,7 @@ class Question < ActiveRecord::Base
 
   has_many :answers
   has_many :votes, as: :votable
+  has_many :reactions, as: :reactable
 
   has_many :voters, through: :votes, source: :user
 
@@ -47,6 +50,7 @@ class Answer < ActiveRecord::Base
   belongs_to :author, class_name: 'User'
 
   has_many :votes, as: :votable
+  has_many :reactions, as: :reactable
 
   has_many :voters, through: :votes, source: :user
 
@@ -58,6 +62,11 @@ class TrackingPixel < ActiveRecord::Base
 end
 
 class Vote < ActiveRecord::Base
+  belongs_to :votable, polymorphic: true
+  belongs_to :user
+end
+
+class Reaction < ActiveRecord::Base
   belongs_to :votable, polymorphic: true
   belongs_to :user
 end