+++ test/db/seeds.rb
@@ -32,11 +32,15 @@ question = forums.first.questions.create!(
 
 question.votes.create!(user: User.find_by!(login: 'alice'))
 
-question.answers.create!(
+answer = question.answers.create!(
   text:   'Oh please go on hack it, ROFL)))',
   best:   true,
   author: User.find_by!(login: 'alice'),
 )
+
+answer.reactions.create!(users.map.with_index do |user, i|
+  { user: user, reaction: ':+1:', created_at: Time.current - 1.hour + i.minutes }
+end)
 
 answer = question.answers.create!(
   text:   'Please, stop',
@@ -55,9 +59,10 @@ question_attrs = %w[first second third fourth fifth].map { |name| { name: name, 
 Question.create!(question_attrs)
 
 Profile.create!([
-  { user: users[0], name: "Profile for user 0", title: "Title for user 0" },
-  { user: users[1], name: "Profile for user 1", title: "Title for user 1" },
-  { user: users[2], name: "Profile for user 2", title: "Title for user 2" },
+  { user: users[0], default: true,  name: "Default profile for user 0", title: "Default title for user 0" },
+  { user: users[0], default: false, name: "Profile for user 0", title: "Title for user 0" },
+  { user: users[1], default: true,  name: "Profile for user 1", title: "Title for user 1" },
+  { user: users[2], default: true,  name: "Profile for user 2", title: "Title for user 2" },
 ])
 
 Version.create!([